refactor: extract component into shared
