+++ src/components/shared/InvisibleCopyableAddress/index.ts
@@ -1,1 +1,2 @@
 export { default } from './InvisibleCopyableAddress';
+export { default as InvisibleCopyableMaskedAddress } from './InvisibleCopyableMaskedAddress';
